{{cookiecutter.project.slug}}/requirements/base.txt
{% if cookiecutter.project.settings.version == '3.1.x' %}Django==3.1.5{% else %}django==2.2.17{% endif %}
django-crispy-forms==1.10.0
django-environ==0.4.5
django-extensions==2.2.9
pip==20.3.4
Pillow==8.1.0
psycopg2-binary
pytz
six==1.15.0
sorl-thumbnail==12.7.0
django-compressor==2.4
django-libsass==0.8
django-meta==2.0.0